Website Changes
I have changed the layout of the Agile South West Website to how we
wanted it. I then changed the design to something more better. Finally I
changed the background to sticky notes placed at an angle. The
background wil have to be re-done as I have got the image from the
Internet but I presume we will do that next time we meet.